Example DAT code for the vtk package
This adds a variable loader for .vtk datasets (that can be loaded by the
vtkDataSetReader module).

A single plot is provided from a subworkflow, rendering a dataset with
vtkRenderer.

Two operations are provided, for appending two datasets
(vtkAppendFilter), computing the outline (vtkOutlineFilter) and cutting
along a plane (actually uses a PythonSource).