Example DAT code for the matplotlib package
Most of the stuff that was initially developed here for DAT was moved to
the tabledata package.

For tabledata, this adds variable loaders for binary data
(numpy.fromfile), numpy array (numpy.load) and CSV. Also adds
multiplication by a scalar and data conversion operations.

For matplotlib, this adds 7 plots defined from subworkflows.